Adding ruby 1.9 support back in and conditionally switching. Also reverted rspec testing in favor of Test::Unit
