+++ Gemfile
@@ -5,6 +5,10 @@ gemspec
 group :development do
   gem "rake"
   gem "pry-nav"
+  gem 'unicode'
+  gem 'unf'
+  # gem 'unicode_utils'
+  gem 'activesupport'
 end
 
 group :test do
