+++ lib/eprun/version.rb
@@ -1,9 +1,9 @@
-# coding: utf-8
+# encoding: utf-8
 
 # Copyright 2010-2013 Ayumu Nojima (野島 歩) and Martin J. Dürst (duerst@it.aoyama.ac.jp)
 # available under the same licence as Ruby itself
 # (see http://www.ruby-lang.org/en/LICENSE.txt)
 
-module Eprun
+class Eprun
   VERSION = "1.1.0"
 end
