+++ tasks/tables.rb.erb
@@ -1,5 +1,9 @@
 # encoding: utf-8
 
+# Copyright 2010-2013 Ayumu Nojima (野島 歩) and Martin J. Dürst (duerst@it.aoyama.ac.jp)
+# available under the same licence as Ruby itself
+# (see http://www.ruby-lang.org/en/LICENSE.txt)
+
 # automatically generated by rake generate_tables
 
 module Eprun
